Improved CTCSS squelch. Added a signal strength detector that can be used with the voter. Added a de-emphasis filter. It is now possible to specify fractions for the CTCSS_FQ configuration variable.
git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@861 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/remoterx/NetUplink.h
@@ -180,7 +180,7 @@ class NetUplink : public Uplink
      * @brief 	Pass on detected tone
      * @param 	tone_fq The frequency of the received tone
      */
-    void toneDetected(int tone_fq);
+    void toneDetected(float tone_fq);
     
     /**
      * @brief 	Pass on received audio
